// types/menopause.ts
// Define el estado de la lista de verificación de autocuidado en la menopausia
export interface MenopauseChecklist {
  calciumIntake: boolean; // Ingesta o receta de calcio natural
  exerciseActive: boolean; // Ejercicio de fuerza y paso ligero
  mammogramDone: boolean; // Mamografía anual preventiva
  pelvicFloorTrainting: boolean; // Ejercicios de Kegel diarios
  heartCheck: boolean; // Chequeo de presión arterial e hipertensión
}